Ruby 2.3 Workaround
There appears to be a problem in Ruby 2.3 around upgrading the supports?
method from protected to private. This does not work, for certain ordering of
the tests (but not all) so this change works around the bug by redefining the
method rather than upgrading it. This may be revisited with future releases of
Ruby.