Generic Services
Previously in order to represent a service as a JNDI resource, direct
knowledge of the service type was required.  This allowed a very tight
integration, but was brittle in that most services were not know about.  This
change adds support for any unknown service to be converted into a resource.

[resolves #33]